+++ osmdroid-android/src/org/osmdroid/tileprovider/modules/DatabaseFileArchive.java
@@ -32,10 +32,10 @@ public class DatabaseFileArchive implements IArchiveFile {
 		try {
 			InputStream ret = null;
 			final String[] tile = {"tile"};
-			final int x = pTile.getX();
-			final int y = pTile.getY();
-			final int z = pTile.getZoomLevel();
-			final int index = ((z << z) + x << z) + y;
+			final long x = (long) pTile.getX();
+			final long y = (long) pTile.getY();
+			final long z = (long) pTile.getZoomLevel();
+			final long index = ((z << z) + x << z) + y;
 			final Cursor cur = mDatabase.query("tiles", tile, "key = " + index + " and provider = '" + pTileSource.name() + "'", null, null, null, null);
 			if(cur.getCount() != 0) {
 				cur.moveToFirst();
